Separate mocks function
